+++ CellFactory.go
@@ -13,7 +13,7 @@ package ui
 type CellFactory interface {
 	// CellHeight returns the height to use for the cells. A value less than 1 indicates that each
 	// cell's height may be different.
-	CellHeight() float32
+	CellHeight() float64
 
 	// CreateCell creates a new cell for 'owner' using 'element' as the content. 'index' indicates
 	// which row the element came from. 'selected' indicates the cell should be created in its
