+++ font/Font.go
@@ -19,7 +19,7 @@ import (
 
 const (
 	// PangoScale is a constant used by the Pango font engine.
-	PangoScale = float32(C.PANGO_SCALE)
+	PangoScale = float64(C.PANGO_SCALE)
 )
 
 var (
@@ -30,9 +30,9 @@ var (
 // Font represents a font.
 type Font struct {
 	pfd           *C.PangoFontDescription
-	leading       float32
-	ascent        float32
-	descent       float32
+	leading       float64
+	ascent        float64
+	descent       float64
 	monospaced    uint8
 	metricsLoaded bool
 }
@@ -148,12 +148,12 @@ func (d *Font) SetStretch(stretch Stretch) {
 }
 
 // Size the size of the font, in points.
-func (d *Font) Size() float32 {
-	return float32(C.pango_font_description_get_size(d.pfd)) / PangoScale
+func (d *Font) Size() float64 {
+	return float64(C.pango_font_description_get_size(d.pfd)) / PangoScale
 }
 
 // SetSize sets the size of the font, in points.
-func (d *Font) SetSize(size float32) {
+func (d *Font) SetSize(size float64) {
 	C.pango_font_description_set_size(d.pfd, C.gint(size*PangoScale))
 	d.metricsLoaded = false
 }
@@ -172,25 +172,25 @@ func (d *Font) Monospaced() bool {
 }
 
 // Leading returns the amount of space before the ascent.
-func (d *Font) Leading() float32 {
+func (d *Font) Leading() float64 {
 	d.loadMetrics()
 	return d.leading
 }
 
 // Ascent returns the amount of space used from the baseline to the top of the tallest character.
-func (d *Font) Ascent() float32 {
+func (d *Font) Ascent() float64 {
 	d.loadMetrics()
 	return d.ascent
 }
 
 // Descent returns the amount of space used from the baseline to the bottom.
-func (d *Font) Descent() float32 {
+func (d *Font) Descent() float64 {
 	d.loadMetrics()
 	return d.descent
 }
 
 // Height returns the overall height of the font, effectively Leading() + Ascent() + Descent().
-func (d *Font) Height() float32 {
+func (d *Font) Height() float64 {
 	d.loadMetrics()
 	return d.leading + d.ascent + d.descent
 }
@@ -200,8 +200,8 @@ func (d *Font) loadMetrics() {
 		d.metricsLoaded = true
 		if f := C.pango_font_map_load_font(C.pango_cairo_font_map_get_default(), context, d.pfd); f != nil {
 			metrics := C.pango_font_get_metrics(f, C.pango_language_get_default())
-			d.ascent = float32(C.pango_font_metrics_get_ascent(metrics)) / PangoScale
-			d.descent = float32(C.pango_font_metrics_get_descent(metrics)) / PangoScale
+			d.ascent = float64(C.pango_font_metrics_get_ascent(metrics)) / PangoScale
+			d.descent = float64(C.pango_font_metrics_get_descent(metrics)) / PangoScale
 			d.leading = d.descent / 2
 			C.pango_font_metrics_unref(metrics)
 		} else {
@@ -229,12 +229,12 @@ func (d *Font) Measure(text string) geom.Size {
 	var width, height C.int
 	C.pango_layout_get_size(layout, &width, &height)
 	d.destroyLayout(layout, gc)
-	return geom.Size{Width: float32(width) / PangoScale, Height: d.Leading() + float32(height)/PangoScale}
+	return geom.Size{Width: float64(width) / PangoScale, Height: d.Leading() + float64(height)/PangoScale}
 }
 
 // IndexForPosition returns the rune index within the string for the specified x-coordinate, where
 // 0 is the start of the string.
-func (d *Font) IndexForPosition(x float32, text string) int {
+func (d *Font) IndexForPosition(x float64, text string) int {
 	layout, gc := d.createLayout(text)
 	var index, trailing C.int
 	C.pango_layout_xy_to_index(layout, C.int(x*PangoScale), 0, &index, &trailing)
@@ -244,12 +244,12 @@ func (d *Font) IndexForPosition(x float32, text string) int {
 
 // PositionForIndex returns the x-coordinate where the specified rune index starts. The returned
 // coordinate assumes 0 is the start of the string.
-func (d *Font) PositionForIndex(index int, text string) float32 {
+func (d *Font) PositionForIndex(index int, text string) float64 {
 	layout, gc := d.createLayout(text)
 	var x C.int
 	C.pango_layout_index_to_line_x(layout, C.int(index), 0, nil, &x)
 	d.destroyLayout(layout, gc)
-	return float32(x) / PangoScale
+	return float64(x) / PangoScale
 }
 
 func (d *Font) createLayout(text string) (layout *C.PangoLayout, gc *C.cairo_t) {
